feat: pattern, wheel entity 생성 #3
