+++ src/main/java/com/endlesshorses/oot/custom/color/entity/Color.java
@@ -1,0 +1,22 @@
+package com.endlesshorses.oot.custom.color.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Color {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+	@Column(unique = true, nullable = false)
+	private String name;
+	@Column(nullable = false)
+	private Long red;
+	@Column(nullable = false)
+	private Long green;
+	@Column(nullable = false)
+	private Long blue;
+}
