+++ src/main/java/com/endlesshorses/oot/custom/font/entity/Font.java
@@ -1,0 +1,20 @@
+package com.endlesshorses.oot.custom.font.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Font {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+	@Column(unique = true, nullable = false)
+	private String name;
+	@Column(nullable = false)
+	private Long price;
+	@Column(nullable = false)
+	private String imageUrl;
+}
