+++ src/main/java/com/endlesshorses/oot/custom/wheel/entity/Wheel.java
@@ -1,0 +1,22 @@
+package com.endlesshorses.oot.custom.wheel.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Wheel {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+	@Column(unique = true, nullable = false)
+	private String name;
+	@Column(nullable = false)
+	private Long price;
+	@Column(nullable = false)
+	private String imageUrl;
+	@Column(nullable = false)
+	private String explanation;
+}
